Attempt to address #324 - merge race condition
Merge SegmentEntry are not requested atomically.
Because of that, the list of entries may change in the middle of
`start_merge` executing.

The `segment_entries` fetched for each of the segment id end
up having a different revision. As a result, we end up protecting a
`.del` files  associated to different opstamp.

We select the delete queue opstamp by looking at the first segment entry
only.

If GC also kicks in, the .del files required to catch up with the
delete queue may be deleted.